# internal/shell/http/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Scheduler API",
    "description": "A job scheduling service that allows users to create, manage, and execute scheduled jobs.\n\n## Authentication\nAll endpoints require Red Hat identity authentication via the `X-Rh-Identity` header.\n\n## Schedule Format\nJobs can be scheduled using standard 5-field cron expressions:\n- Format: `minute hour day month weekday`\n- All schedules are interpreted in UTC timezone\n",
    "version": "1.0.0",
    "contact": {
      "name": "Red Hat"
    }
  },
  "servers": [
    {
      "url": "https://console.redhat.com/api/scheduler/v1",
      "description": "Production server"
    }
  ],
...
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
  "paths": {
...
    "/jobs/{id}/runs": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "Job ID",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "tags": [
          "Job Runs"
        ],
        "summary": "List all runs for a job",
        "description": "Retrieve all execution runs for a specific job. Returns runs in reverse chronological order (most recent first), with pagination support.",
        "operationId": "getJobRuns",
        "parameters": [
          {
            "name": "offset",
            "in": "query",
            "description": "Number of items to skip (for pagination)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "example": 0
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of items to return (max 100)",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            },
            "example": 20
          },
          {
            "name": "sort_by",
            "in": "query",
            "description": "Sort runs by a field, in the form 'field:direction'. Direction is optional and defaults to 'asc'. Allowed fields: start_time, end_time, status, created_at. Defaults to 'start_time:desc'.",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(start_time|end_time|status|created_at)(:(asc|desc))?$",
              "default": "start_time:desc"
            },
            "example": "start_time:desc"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of job runs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedJobRunsResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/jobs/{id}/runs/{run_id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "description": "Job ID",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        },
        {
          "name": "run_id",
          "in": "path",
          "description": "Job run ID",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ],
      "get": {
        "tags": [
          "Job Runs"
        ],
        "summary": "Get a specific job run",
        "description": "Retrieve details of a specific job execution run.",
        "operationId": "getJobRun",
        "responses": {
          "200": {
            "description": "Job run details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRun"
                },
                "examples": {
                  "exportJobRun": {
                    "summary": "Completed export job run",
                    "value": {
                      "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                      "job_id": "550e8400-e29b-41d4-a716-446655440000",
                      "status": "completed",
                      "start_time": "2025-11-07T10:00:00Z",
                      "end_time": "2025-11-07T10:05:30Z",
                      "result_type": "export",
                      "result": {
                        "export_id": "abc-123-def-456",
                        "url": "https://console.redhat.com/api/export/v1/exports/abc-123-def-456"
                      }
                    }
                  },
                  "failedJobRun": {
                    "summary": "Failed job run",
                    "value": {
                      "id": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
                      "job_id": "550e8400-e29b-41d4-a716-446655440000",
                      "status": "failed",
                      "start_time": "2025-11-07T10:00:00Z",
                      "end_time": "2025-11-07T10:00:15Z",
                      "error_message": "Connection timeout"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Rh-Identity",
        "description": "Base64-encoded JSON containing Red Hat identity information."
      }
    },
    "schemas": {
...
      "JobRun": {
        "type": "object",
        "description": "A record of a single job execution",
        "required": [
          "id",
          "job_id",
          "status",
          "start_time"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Unique identifier for the job run",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          },
          "job_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the parent job",
            "example": "550e8400-e29b-41d4-a716-446655440000"
          },
          "job_name": {
            "type": "string",
            "description": "Name of the parent job",
            "example": "Daily export job"
          },
          "status": {
            "$ref": "#/components/schemas/JobRunStatus"
          },
          "start_time": {
            "type": "string",
            "format": "date-time",
            "description": "When the job run started",
            "example": "2025-11-07T10:00:00Z"
          },
          "end_time": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "When the job run completed or failed",
            "example": "2025-11-07T10:05:30Z"
          },
          "error_message": {
            "type": "string",
            "nullable": true,
            "description": "Error message if the run failed",
            "example": "Connection timeout"
          },
          "result_type": {
            "type": "string",
            "description": "Type discriminator for the result object",
            "nullable": true,
            "enum": [
              "export"
            ],
            "example": "export"
          },
          "result": {
            "nullable": true,
            "description": "Result object from the export job execution",
            "allOf": [
              {
                "$ref": "#/components/schemas/ExportResult"
              }
            ]
          }
        }
      },
      "ExportResult": {
        "type": "object",
        "description": "Result from an export job execution. The result_type field at the JobRun level will be 'export'.",
        "required": [
          "export_id"
        ],
        "properties": {
          "export_id": {
            "type": "string",
            "description": "ID of the created export",
            "example": "abc-123-def-456"
          },
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Download URL for the completed export",
            "example": "https://console.redhat.com/api/export/v1/exports/abc-123-def-456"
          }
        }
      },
      "JobRunStatus": {
        "type": "string",
        "description": "Status of a job run",
        "enum": [
          "running",
          "completed",
          "failed"
        ],
        "example": "completed"
      },
...
      "ErrorObject": {
        "type": "object",
        "description": "JSON:API error object",
        "required": [
          "status",
          "title",
          "detail"
        ],
        "properties": {
          "status": {
            "type": "string",
            "description": "HTTP status code as a string",
            "example": "404"
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary of the problem",
            "example": "Job Not Found"
          },
          "detail": {
            "type": "string",
            "description": "Human-readable explanation specific to this occurrence",
            "example": "The job with ID '550e8400-e29b-41d4-a716-446655440000' could not be found or you do not have permission to access it"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "description": "JSON:API error response",
        "required": [
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "description": "Array of error objects",
            "items": {
              "$ref": "#/components/schemas/ErrorObject"
            },
            "minItems": 1
          }
        }
      },
      "PaginationMeta": {
        "type": "object",
        "description": "Pagination metadata",
        "required": [
          "count"
        ],
        "properties": {
          "count": {
            "type": "integer",
            "description": "Total number of items across all pages",
            "example": 100
          }
        }
      },
      "PaginationLinks": {
        "type": "object",
        "description": "Navigation links for pagination",
        "properties": {
          "first": {
            "type": "string",
            "format": "uri",
            "description": "Link to the first page",
            "example": "/api/scheduler/v1/jobs?limit=20&offset=0"
          },
          "last": {
            "type": "string",
            "format": "uri",
            "description": "Link to the last page",
            "example": "/api/scheduler/v1/jobs?limit=20&offset=80"
          },
          "next": {
            "type": "string",
            "format": "uri",
            "description": "Link to the next page (if available)",
            "example": "/api/scheduler/v1/jobs?limit=20&offset=20"
          },
          "prev": {
            "type": "string",
            "format": "uri",
            "description": "Link to the previous page (if available)",
            "example": "/api/scheduler/v1/jobs?limit=20&offset=0"
          }
        }
      },
...
      "PaginatedJobRunsResponse": {
        "type": "object",
        "description": "Paginated response for job runs list",
        "required": [
          "meta",
          "links",
          "data"
        ],
        "properties": {
          "meta": {
            "$ref": "#/components/schemas/PaginationMeta"
          },
          "links": {
            "$ref": "#/components/schemas/PaginationLinks"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JobRun"
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad request - Invalid input or missing required fields",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "examples": {
              "invalidJSON": {
                "summary": "Invalid JSON",
                "value": {
                  "errors": [
                    {
                      "status": "400",
                      "title": "Invalid JSON",
                      "detail": "The request body contains invalid JSON"
                    }
                  ]
                }
              },
              "missingFields": {
                "summary": "Missing required fields",
                "value": {
                  "errors": [
                    {
                      "status": "400",
                      "title": "Missing Required Fields",
                      "detail": "Missing Required Fields"
                    }
                  ]
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "errors": [
                {
                  "status": "404",
                  "title": "Job Not Found",
                  "detail": "The job with ID '550e8400-e29b-41d4-a716-446655440000' could not be found or you do not have permission to access it"
                }
              ]
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            },
            "example": {
              "errors": [
                {
                  "status": "500",
                  "title": "Internal Server Error",
                  "detail": "An unexpected error occurred while processing your request"
                }
              ]
            }
          }
        }
      }
    }
  }
}